apps/lamin/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.6.15
charset-normalizer==2.1.0
cssutils==2.5.0
dataclasses-json==0.5.13
dateparser==1.1.1
feedparser==6.0.10
frozenlist==1.4.0
fsspec==2023.6.0
html2text==2020.1.16
htmldate==1.2.3
idna==3.3
langchain==0.0.245
langsmith==0.0.15
llama-index==0.7.11
lxml==4.9.1
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
neo4j==5.10.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pydantic==1.10.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
regex==2022.3.2
requests==2.28.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2022.1
tzlocal==4.2
Unidecode==1.3.4
urllib3==1.26.10
yarl==1.9.2
